+++ .github/workflows/e2e-ci.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Install pnpm
         run: |
           corepack enable
-          corepack prepare 
+          corepack prepare
       - name: set store
         run: |
           mkdir ~/.pnpm-store
@@ -45,14 +45,14 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-      - name: Use Node 
+      - name: Use Node
         uses: actions/setup-node@39370e3970a6d050c480ffad4ff0ed4d3fdee5af # v4.1.0
         with:
           node-version-file: '.nvmrc'
       - name: Install pnpm
         run: |
           corepack enable
-          corepack prepare 
+          corepack prepare
       - uses: actions/cache@6849a6489940f00c2f30c0fb92c6274307ccb58a # v4.0.2
         with:
           path: ~/.pnpm-store
@@ -83,7 +83,7 @@ jobs:
   e2e-cli-npm:
     needs: [prepare, build]
     strategy:
-      fail-fast: false      
+      fail-fast: false
       matrix:
         pkg:
           [
@@ -93,7 +93,7 @@ jobs:
             npm9,
             npm10
           ]
-        node: [20, 21]
+        node: [20, 22]
     name: ${{ matrix.pkg }}/ ubuntu-latest / ${{ matrix.node }}
     runs-on: ubuntu-latest
     steps:
@@ -104,7 +104,7 @@ jobs:
       - name: Install pnpm
         run: |
           corepack enable
-          corepack prepare 
+          corepack prepare
       - uses: actions/cache@6849a6489940f00c2f30c0fb92c6274307ccb58a # v4.0.2
         with:
           path: ~/.pnpm-store
@@ -136,8 +136,9 @@ jobs:
           [
             pnpm8,
             pnpm9,
+            pnpm10,
           ]
-        node: [20, 21]
+        node: [20, 22]
     name: ${{ matrix.pkg }}/ ubuntu-latest / ${{ matrix.node }}
     runs-on: ubuntu-latest
     steps:
@@ -176,13 +177,13 @@ jobs:
       fail-fast: false
       matrix:
         pkg:
-          [           
+          [
             yarn1,
             yarn2,
             yarn3,
             yarn4
           ]
-        node: [20, 21]
+        node: [20, 22]
     name: ${{ matrix.pkg }}/ ubuntu-latest / ${{ matrix.node }}
     runs-on: ubuntu-latest
     steps:
@@ -193,7 +194,7 @@ jobs:
       - name: Install pnpm
         run: |
           corepack enable
-          corepack prepare 
+          corepack prepare
       - uses: actions/cache@6849a6489940f00c2f30c0fb92c6274307ccb58a # v4.0.2
         with:
           path: ~/.pnpm-store
@@ -215,4 +216,4 @@ jobs:
         run: pnpm --filter @verdaccio/test-cli-commons build
       - name: Test CLI
         run: NODE_ENV=production pnpm test --filter ...@verdaccio/e2e-cli-${{matrix.pkg}}
-      
+
