+++ e2e/cli/e2e-pnpm-commons/index.ts
@@ -1,0 +1,10 @@
+export { runAudit } from './audit';
+export { runDeprecate } from './deprecate';
+export { runInstall } from './install';
+export { runInfo } from './info';
+export { runPing } from './ping';
+export { runPublish } from './publish';
+export { runSearch } from './search';
+export { runStar } from './star';
+export { runUnpublish } from './unpublish';
+export { runDistTag } from './dist-tags';
