+++ e2e/cli/e2e-pnpm-commons/info.ts
@@ -1,0 +1,32 @@
+import { afterAll, beforeAll, describe, expect, test } from 'vitest';
+
+import { addRegistry, initialSetup } from '@verdaccio/test-cli-commons';
+
+export function runInfo(pnpm) {
+  describe('info a package', () => {
+    let registry;
+
+    beforeAll(async () => {
+      const setup = await initialSetup();
+      registry = setup.registry;
+      await registry.init();
+    });
+
+    test('should run pnpm info json body', async () => {
+      const resp = await pnpm(
+        {},
+        'info',
+        'verdaccio',
+        '--json',
+        ...addRegistry(registry.getRegistryUrl())
+      );
+      const parsedBody = JSON.parse(resp.stdout as string);
+      expect(parsedBody.name).toEqual('verdaccio');
+      expect(parsedBody.dependencies).toBeDefined();
+    });
+
+    afterAll(async () => {
+      registry.stop();
+    });
+  });
+}
