+++ e2e/cli/e2e-pnpm-commons/package.json
@@ -1,12 +1,12 @@
 {
   "private": true,
-  "name": "@verdaccio/e2e-cli-pnpm9",
+  "name": "@verdaccio/e2e-cli-pnpm-common",
   "version": "1.0.2-next-8.0",
   "dependencies": {
     "@verdaccio/test-cli-commons": "workspace:2.0.0-next-8.0",
-    "@verdaccio/e2e-cli-pnpm-common": "workspace:*",
-    "pnpm": "9.13.2"
+    "pnpm": "9.0.0-alpha.10"
   },
+  "main": "index.ts",
   "scripts": {
     "test": "vitest run --testTimeout 50000"
   }
