+++ e2e/cli/e2e-pnpm10/.babelrc
@@ -1,0 +1,3 @@
+{
+  "extends": "../../../.babelrc"
+}
