+++ e2e/cli/e2e-pnpm10/CHANGELOG.md
@@ -1,0 +1,8 @@
+# @verdaccio/e2e-cli-pnpm9
+
+## 1.0.2-next-8.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @verdaccio/test-cli-commons@2.0.0-next-8.0
