+++ e2e/cli/e2e-pnpm10/audit.spec.ts
