+++ e2e/cli/e2e-pnpm10/deprecate.spec.ts
