+++ e2e/cli/e2e-pnpm10/dist-tags.spec.ts
