+++ e2e/cli/e2e-pnpm10/info.spec.ts
