+++ e2e/cli/e2e-pnpm10/install.spec.ts
