+++ e2e/cli/e2e-pnpm10/package.json
@@ -1,11 +1,11 @@
 {
   "private": true,
-  "name": "@verdaccio/e2e-cli-pnpm9",
-  "version": "1.0.2-next-8.0",
+  "name": "@verdaccio/e2e-cli-pnpm10",
+  "version": "1.0.0",
   "dependencies": {
     "@verdaccio/test-cli-commons": "workspace:2.0.0-next-8.0",
     "@verdaccio/e2e-cli-pnpm-common": "workspace:*",
-    "pnpm": "9.13.2"
+    "pnpm": "10.0.0-alpha.2"
   },
   "scripts": {
     "test": "vitest run --testTimeout 50000"
