+++ e2e/cli/e2e-pnpm10/ping.spec.ts
