+++ e2e/cli/e2e-pnpm10/publish.spec.ts
