+++ e2e/cli/e2e-pnpm10/search.spec.ts
