+++ e2e/cli/e2e-pnpm10/star.spec.ts
