+++ e2e/cli/e2e-pnpm10/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../tsconfig.reference.json",
+  "references": [
+    {
+      "path": "../cli-commons"
+    }
+  ]
+}
