+++ e2e/cli/e2e-pnpm10/unpublish.spec.ts
@@ -1,0 +1,9 @@
+import { describe } from 'vitest';
+
+import { runUnpublish } from '@verdaccio/e2e-cli-pnpm-common';
+
+import { pnpm } from './utils';
+
+describe('unpublish a package', () => {
+  runUnpublish(pnpm);
+});
