+++ e2e/cli/e2e-pnpm8/package.json
@@ -4,7 +4,8 @@
   "version": "1.0.2-next-8.0",
   "dependencies": {
     "@verdaccio/test-cli-commons": "workspace:2.0.0-next-8.0",
-    "pnpm": "8.15.5"
+    "@verdaccio/e2e-cli-pnpm-common": "workspace:*",
+    "pnpm": "8.15.9"
   },
   "scripts": {
     "test": "vitest run --testTimeout 50000"
