+++ e2e/cli/e2e-pnpm9/info.spec.ts
@@ -1,32 +1,9 @@
-import { afterAll, beforeAll, describe, expect, test } from 'vitest';
+import { describe } from 'vitest';
 
-import { addRegistry, initialSetup } from '@verdaccio/test-cli-commons';
+import { runInfo } from '@verdaccio/e2e-cli-pnpm-common';
 
 import { pnpm } from './utils';
 
-describe('install a package', () => {
-  let registry;
-
-  beforeAll(async () => {
-    const setup = await initialSetup();
-    registry = setup.registry;
-    await registry.init();
-  });
-
-  test('should run pnpm info json body', async () => {
-    const resp = await pnpm(
-      {},
-      'info',
-      'verdaccio',
-      '--json',
-      ...addRegistry(registry.getRegistryUrl())
-    );
-    const parsedBody = JSON.parse(resp.stdout as string);
-    expect(parsedBody.name).toEqual('verdaccio');
-    expect(parsedBody.dependencies).toBeDefined();
-  });
-
-  afterAll(async () => {
-    registry.stop();
-  });
+describe('info a package', () => {
+  runInfo(pnpm);
 });
