+++ e2e/cli/e2e-pnpm9/install.spec.ts
@@ -1,37 +1,9 @@
-import { afterAll, beforeAll, describe, expect, test } from 'vitest';
+import { describe } from 'vitest';
 
-import { addRegistry, initialSetup, prepareGenericEmptyProject } from '@verdaccio/test-cli-commons';
+import { runInstall } from '@verdaccio/e2e-cli-pnpm-common';
 
 import { pnpm } from './utils';
 
 describe('install a project packages', () => {
-  let registry;
-
-  beforeAll(async () => {
-    const setup = await initialSetup();
-    registry = setup.registry;
-    await registry.init();
-  });
-
-  test('should run npm install json body', async () => {
-    const { tempFolder } = await prepareGenericEmptyProject(
-      'something',
-      '1.0.0-patch',
-      registry.port,
-      registry.getToken(),
-      registry.getRegistryUrl(),
-      { react: '18.2.0' }
-    );
-    const resp = await pnpm(
-      { cwd: tempFolder },
-      'install',
-      '--reporter=default',
-      ...addRegistry(registry.getRegistryUrl())
-    );
-    expect(resp.stdout).toMatch(/react/);
-  });
-
-  afterAll(async () => {
-    registry.stop();
-  });
+  runInstall(pnpm);
 });
