+++ e2e/cli/e2e-pnpm9/publish.spec.ts
@@ -1,42 +1,9 @@
-import { afterAll, beforeAll, describe, expect, test } from 'vitest';
+import { describe } from 'vitest';
 
-import { addRegistry, initialSetup, prepareGenericEmptyProject } from '@verdaccio/test-cli-commons';
+import { runPublish } from '@verdaccio/e2e-cli-pnpm-common';
 
 import { pnpm } from './utils';
 
-describe('install a package', () => {
-  let registry;
-
-  beforeAll(async () => {
-    const setup = await initialSetup();
-    registry = setup.registry;
-    await registry.init();
-  });
-
-  test.each([['verdaccio-memory', 'verdaccio', '@verdaccio/foo', '@verdaccio/some-foo']])(
-    'should publish a package %s',
-    async (pkgName) => {
-      const { tempFolder } = await prepareGenericEmptyProject(
-        pkgName,
-        '1.0.0-patch',
-        registry.port,
-        registry.getToken(),
-        registry.getRegistryUrl()
-      );
-      const resp = await pnpm(
-        { cwd: tempFolder },
-        'publish',
-        '--json',
-        ...addRegistry(registry.getRegistryUrl())
-      );
-      const parsedBody = JSON.parse(resp.stdout as string);
-      expect(parsedBody.name).toEqual(pkgName);
-      expect(parsedBody.files).toBeDefined();
-      expect(parsedBody.files).toBeDefined();
-    }
-  );
-
-  afterAll(async () => {
-    registry.stop();
-  });
+describe('publish a package', () => {
+  runPublish(pnpm);
 });
